moved function to `parametric_images.py`
